Refactor Explorer backend and harden services
