+++ src/Core/bridge/apps/explorer/src/json.rs
@@ -1,0 +1,28 @@
+pub fn array<T>(items: &[T], f: fn(&T) -> String) -> String {
+    let mut out = String::with_capacity(items.len() * 200);
+    out.push('[');
+    for (i, item) in items.iter().enumerate() {
+        if i > 0 {
+            out.push(',');
+        }
+        out.push_str(&f(item));
+    }
+    out.push(']');
+    out
+}
+
+pub fn escape(s: &str) -> String {
+    let mut out = String::with_capacity(s.len() + 8);
+    for c in s.chars() {
+        match c {
+            '\\' => out.push_str("\\\\"),
+            '"' => out.push_str("\\\""),
+            '\n' => out.push_str("\\n"),
+            '\r' => out.push_str("\\r"),
+            '\t' => out.push_str("\\t"),
+            c if (c as u32) < 0x20 => out.push_str(&format!("\\u{:04x}", c as u32)),
+            c => out.push(c),
+        }
+    }
+    out
+}
