+++ src/System/portal/astrea_filechooser_portal.py
@@ -190,20 +190,31 @@ def run_dialog(mode, title, options):
     env["BENCH_FILE_DIALOG_OPTIONS"] = dialog_options
     env["BENCH_FILE_DIALOG_RESULT_FILE"] = result_path
     log_debug(f"run_dialog mode={mode} title={title!r}")
+    if not Path(PORTAL_DIALOG_QML).is_file():
+        log_debug(f"Portal dialog QML not found: {PORTAL_DIALOG_QML}")
+        try:
+            os.unlink(result_path)
+        except FileNotFoundError:
+            pass
+        return {"accepted": False}
 
     child = None
     output_buffer = ""
     qslog_path = None
     try:
-        child = subprocess.Popen(
-            [QS_BIN, "-p", PORTAL_DIALOG_QML],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.STDOUT,
-            env=env,
-            start_new_session=True,
-            text=True,
-            bufsize=1,
-        )
+        try:
+            child = subprocess.Popen(
+                [QS_BIN, "-p", PORTAL_DIALOG_QML],
+                stdout=subprocess.PIPE,
+                stderr=subprocess.STDOUT,
+                env=env,
+                start_new_session=True,
+                text=True,
+                bufsize=1,
+            )
+        except OSError as exc:
+            log_debug(f"Failed to start portal dialog: {exc}")
+            return {"accepted": False}
         selector = selectors.DefaultSelector()
         if child.stdout is not None:
             selector.register(child.stdout, selectors.EVENT_READ)
